Fix circular import issues
