+++ megatron/core/models/hybrid/hybrid_model.py
@@ -8,6 +8,7 @@ from torch import Tensor
 from megatron.core import tensor_parallel
 from megatron.core.config_logger import has_config_logger_enabled, log_config_to_disk
 from megatron.core.inference.contexts import BaseInferenceContext
+from megatron.core.inference.utils import InferenceMode
 from megatron.core.models.common.embeddings.language_model_embedding import LanguageModelEmbedding
 from megatron.core.models.common.embeddings.rotary_pos_embedding import RotaryEmbedding
 from megatron.core.models.common.embeddings.yarn_rotary_pos_embedding import YarnRotaryEmbedding
@@ -347,7 +348,7 @@ class HybridModel(LanguageModule, GraphableMegatronModule):
         Check if we should call the local cudagraph path.
         """
         if (
-            BaseInferenceContext.is_active()
+            InferenceMode.is_active()
             and hasattr(self, 'cudagraph_manager')
             and (
                 kwargs.get('inference_context') is not None
@@ -407,7 +408,7 @@ class HybridModel(LanguageModule, GraphableMegatronModule):
 
         inference_context = deprecate_inference_params(inference_context, inference_params)
 
-        in_inference_mode = BaseInferenceContext.is_active()
+        in_inference_mode = InferenceMode.is_active()
 
         if in_inference_mode:
             assert runtime_gather_output, "Inference must always gather TP logits"
