+++ megatron/core/ssm/mamba_mixer.py
@@ -23,6 +23,7 @@ from megatron.core.inference.contexts.attention_context.triton.tensor_ops import
     tensor_masked_update,
     tensor_merge,
 )
+from megatron.core.inference.utils import InferenceMode
 from megatron.core.packed_seq_params import PackedSeqParams
 from megatron.core.process_groups_config import ProcessGroupCollection
 from megatron.core.ssm.ops.causal_conv1d_triton import causal_conv1d_update
@@ -427,7 +428,7 @@ class MambaMixer(MegatronModule):
 
         inference_context = deprecate_inference_params(inference_context, inference_params)
 
-        in_inference_mode = BaseInferenceContext.is_active()
+        in_inference_mode = InferenceMode.is_active()
 
         _, batch, dim = hidden_states.shape
         conv_state, ssm_state = None, None
