+++ megatron/core/transformer/moe/router.py
@@ -5,7 +5,7 @@ from typing import Optional, Union
 
 import torch
 
-from megatron.core.inference.contexts import BaseInferenceContext
+from megatron.core.inference.utils import InferenceMode
 from megatron.core.jit import jit_fuser
 from megatron.core.transformer.module import MegatronModule
 from megatron.core.transformer.moe.moe_utils import (
@@ -816,7 +816,7 @@ class InferenceTopKRouter(TopKRouter):
                 - top_indices: Selected expert indices [num_tokens, topk]
         """
 
-        if not BaseInferenceContext.is_active():
+        if not InferenceMode.is_active():
             return super().forward(input, padding_mask)
 
         return self._forward(input, padding_mask)
